refactor(eval): Move get_model to utils
