+++ syntheseus/reaction_prediction/cli/eval.py
@@ -22,7 +22,7 @@ from dataclasses import dataclass, field, fields
 from functools import partial
 from itertools import islice
 from statistics import mean, median
-from typing import Any, Callable, Dict, Generic, Iterable, List, Optional, Set, Union, cast
+from typing import Any, Callable, Dict, Generic, Iterable, List, Optional, Set, cast
 
 import numpy as np
 from more_itertools import batched
@@ -52,6 +52,7 @@ from syntheseus.reaction_prediction.utils.metrics import (
     compute_total_time,
 )
 from syntheseus.reaction_prediction.utils.misc import asdict_extended, set_random_seed
+from syntheseus.reaction_prediction.utils.model_loading import get_model
 
 logger = logging.getLogger(__file__)
 
@@ -439,35 +440,6 @@ def print_and_save(results: EvalResults, config: EvalConfig, suffix: str = "") -
 
         with open(outfile, "w") as outfile_stream:
             outfile_stream.write(json.dumps(results_dict))
-
-
-def get_model(
-    config: Union[BackwardModelConfig, ForwardModelConfig], batch_size: int, num_gpus: int
-) -> ReactionModel:
-    def model_fn(device):
-        return config.model_class.value(
-            model_dir=config.model_dir, device=device, **config.model_kwargs
-        )
-
-    if num_gpus == 0:
-        return model_fn("cpu")
-    elif num_gpus == 1:
-        return model_fn("cuda:0")
-    else:
-        if batch_size < num_gpus:
-            raise ValueError(f"Cannot split batch of size {batch_size} across {num_gpus} GPUs")
-
-        batch_size_per_gpu = batch_size // num_gpus
-
-        if batch_size_per_gpu < 16:
-            logger.warning(f"Batch size per GPU is very small: ~{batch_size_per_gpu}")
-
-        try:
-            from syntheseus.reaction_prediction.utils.parallel import ParallelReactionModel
-        except ModuleNotFoundError:
-            raise ValueError("Multi-GPU evaluation is only supported for torch-based models")
-
-        return ParallelReactionModel(model_fn, devices=[f"cuda:{idx}" for idx in range(num_gpus)])
 
 
 def run_from_config(
